+++ internal/data/database.go
@@ -1,4 +1,4 @@
-package persistence
+package data
 
 import (
 	"fmt"
